+++ eras/conway/impl/src/Cardano/Ledger/Conway/Rules/GovCert.hs
@@ -259,6 +259,7 @@ conwayGovCertTransition = do
                     )
                     mAnchor
                     ppDRepDeposit
+                    mempty
                 )
                 vsDReps
           }